Condensed channel commands into one command
